requirements.txt
openai
PyPDF2
numpy
tiktoken
uproot
awkward
pdfminer.six
scikit-learn
